# docs/swagger.yaml
components:
  schemas:
...
    Atlas:
      properties:
        primary_index:
          type: integer
        trees:
          additionalProperties:
            items:
              type: integer
            type: array
          type: object
        user_id:
          type: integer
      required:
      - primary_index
      - trees
      - user_id
      type: object
...
    BplCharacter:
      properties:
        ascendancy:
          type: string
        ascendancy_points:
          type: integer
        atlas_node_count:
          type: integer
        event_id:
          type: integer
        id:
          type: string
        level:
          type: integer
        main_skill:
          type: string
        name:
          type: string
        pantheon:
          type: boolean
        user_id:
          type: integer
        void_stones:
          type: integer
      required:
      - ascendancy
      - ascendancy_points
      - atlas_node_count
      - event_id
      - id
      - level
      - main_skill
      - name
      - pantheon
      - void_stones
      type: object
...
    Event:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        teams:
          items:
            $ref: '#/components/schemas/Team'
          type: array
          uniqueItems: false
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - id
      - is_current
      - is_locked
      - is_main_event
      - is_public
      - max_size
      - name
      - teams
      - uses_medals
      - waitlist_size
      type: object
...
    LadderEntry:
      properties:
        armour:
          type: integer
        ascendancy:
          type: string
        ascendancy_points:
          type: integer
        atlas_points:
          type: integer
        attack_block:
          type: integer
        character_id:
          type: string
        character_name:
          type: string
        delve_depth:
          type: integer
        discord_id:
          type: string
        discord_name:
          type: string
        dps:
          type: integer
        ehp:
          type: integer
        ele_max_hit:
          type: integer
        es:
          type: integer
        evasion:
          type: integer
        high_level_flasks:
          type: integer
        hp:
          type: integer
        item_indexes:
          items:
            type: integer
          type: array
          uniqueItems: false
        last_active:
          type: integer
        level:
          type: integer
        lowest_ele_res:
          type: integer
        main_skill:
          type: string
        mana:
          type: integer
        movement_speed:
          type: integer
        pantheon:
          type: boolean
        phys_max_hit:
          type: integer
        poe_account:
          type: string
        rank:
          type: integer
        spell_block:
          type: integer
        team_id:
          type: integer
        twitch_name:
          type: string
        user_id:
          type: integer
        voidstones:
          type: integer
        xp:
          type: integer
      required:
      - armour
      - ascendancy
      - ascendancy_points
      - atlas_points
      - attack_block
      - character_id
      - character_name
      - delve_depth
      - discord_id
      - discord_name
      - dps
      - ehp
      - ele_max_hit
      - es
      - evasion
      - high_level_flasks
      - hp
      - item_indexes
      - last_active
      - level
      - lowest_ele_res
      - main_skill
      - mana
      - movement_speed
      - pantheon
      - phys_max_hit
      - poe_account
      - rank
      - spell_block
      - team_id
      - user_id
      - voidstones
      - xp
      type: object
...
    Team:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - event_id
      - id
      - name
      type: object
...
    GameVersion:
      enum:
      - poe1
      - poe2
      type: string
      x-enum-varnames:
      - PoE1
      - PoE2
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
...
  /events/{event_id}:
    delete:
      description: Deletes an event
      operationId: DeleteEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - event
    get:
      description: Fetches an event by id
      operationId: GetEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...
  /events/{event_id}/characters:
    get:
      description: Get all characters for an event
      operationId: GetCharactersForEvent
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BplCharacter'
                type: array
          description: OK
      tags:
      - characters
...
  /events/{event_id}/ladder:
    get:
      description: Get the ladder for an event
      operationId: GetLadder
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: only show ladder entries from this timestamp after event start
        in: query
        name: hours_after_event_start
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/LadderEntry'
                type: array
          description: OK
      tags:
      - ladder
...
  /events/{event_id}/team/{team_id}/atlas:
    get:
      description: Get atlas trees for your team for an event
      operationId: GetTeamAtlasesForEvent
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: Team ID
        in: path
        name: team_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Atlas'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - atlas
...